# Improve performance of `derivative` with FiniteDifferences and ForwardDiff