--- a/ext/AbstractDifferentiationFiniteDifferencesExt.jl
+++ b/ext/AbstractDifferentiationFiniteDifferencesExt.jl
@@ -32,4 +32,9 @@ function AD.pullback_function(ba::AD.FiniteDifferencesBackend, f, xs...)
     end
 end
 
+# Better performance: issue #87
+function AD.derivative(ba::AD.FiniteDifferencesBackend, f::TF, x::Real) where {TF<:Function}
+    return (ba.method(f, x),)
+end
+
 end # module