Fixes #87. With this PR, the benchmark in #87 yields on my computer
```julia
julia> @benchmark with_AD(1.)
BenchmarkTools.Trial: 10000 samples with 960 evaluations.
 Range (min … max):  87.060 ns …  1.505 μs  ┊ GC (min … max): 0.00% … 88.90%
 Time  (median):     89.291 ns              ┊ GC (median):    0.00%
 Time  (mean ± σ):   92.532 ns ± 30.948 ns  ┊ GC (mean ± σ):  0.75% ±  2.17%

  █▆▆▆▃▂ ▄▅▄▂▁▁▁ ▄▂▁▁                                       ▁ ▂
  ███████████████████▇▆▇▇█▇▇▇▇▆▅▆▅▅▅▅▄▄▃▃▄▄▁▆▃▁▃▄▃▃▅▃▃▁▁▄█▆▄█ █
  87.1 ns      Histogram: log(frequency) by time       138 ns <

 Memory estimate: 32 bytes, allocs estimate: 2.

julia> @benchmark without_AD(1.)
BenchmarkTools.Trial: 10000 samples with 961 evaluations.
 Range (min … max):  86.429 ns …  1.699 μs  ┊ GC (min … max): 0.00% … 91.77%
 Time  (median):     89.110 ns              ┊ GC (median):    0.00%
 Time  (mean ± σ):   91.506 ns ± 35.367 ns  ┊ GC (mean ± σ):  0.90% ±  2.23%

  ▃█▇▃▃██▄▄▂▁   ▂▃▄▄▄▂▁▁       ▃▂                             ▂
  ████████████▇██████████▇██▇▇████▇███▆▆▅▂▅▄▅▇█████▅▅▆▆▅▆▅▂▃▆ █
  86.4 ns      Histogram: log(frequency) by time       113 ns <

 Memory estimate: 32 bytes, allocs estimate: 2.
```